+++ src/Utils/Directory.php
@@ -3,52 +3,40 @@
 namespace duncan3dc\Sonos\Utils;
 
 use duncan3dc\Sonos\Interfaces\Utils\DirectoryInterface;
-use League\Flysystem\Adapter\Local;
 use League\Flysystem\Filesystem;
-use League\Flysystem\FilesystemInterface;
+use League\Flysystem\FilesystemOperator;
+use League\Flysystem\Local\LocalFilesystemAdapter;
 
 /**
  * Represents a shared directory.
  */
 final class Directory implements DirectoryInterface
 {
-    /**
-     * @var FilesystemInterface $filesystem The full path to the share on the local filesystem.
-     */
-    private $filesystem;
+    private FilesystemOperator $filesystem;
 
     /**
      * @var string $share The full path to the share (including the hostname).
      */
-    private $share;
+    private string $share;
 
     /**
      * @var string $directory The name of the directory (to be appended to both $filesystem and $share).
      */
-    private $directory;
+    private string $directory;
 
 
     /**
      * Create a Directory instance to represent a file share.
      *
-     * @param FilesystemInterface|string $filesystem The full path to the share on the local filesystem.
+     * @param FilesystemOperator|string $filesystem The full path to the share on the local filesystem.
      * @param string $share The full path to the share (including the hostname).
      * @param string $directory The name of the directory (to be appended to both $filesystem and $share).
      */
-    public function __construct($filesystem, string $share, string $directory)
+    public function __construct(FilesystemOperator|string $filesystem, string $share, string $directory)
     {
-        # If a string was passed then convert it to a Filesystem instance
         if (is_string($filesystem)) {
-            $adapter = new Local($filesystem);
+            $adapter = new LocalFilesystemAdapter($filesystem);
             $filesystem = new Filesystem($adapter);
-        }
-
-        # Ensure we got a Filesystem instance
-        if (!$filesystem instanceof FilesystemInterface) {
-            $error = "Invalid filesystem,";
-            $error .= " must be an instance of " . FilesystemInterface::class;
-            $error .= " or a string containing a local path";
-            throw new \InvalidArgumentException($error);
         }
 
         $this->filesystem = $filesystem;
@@ -77,7 +65,7 @@ final class Directory implements DirectoryInterface
      */
     public function has(string $file): bool
     {
-        return $this->filesystem->has("{$this->directory}/{$file}");
+        return $this->filesystem->fileExists("{$this->directory}/{$file}");
     }