+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 ### Changed
 
 * [Logging] setLogger() no longer returns anything (to match the psr/log interface).
+* [Dependencies] Switched to league/flysystem v3 ([#132](https://github.com/duncan3dc/sonos/issues/132)).
 * [Support] Dropped support for PHP 7.3, 7.4, and 8.0.
 
 --------